chore(model)!: Also move `SnippetChoices` under `snippet`
This is a follow-up to cbd4851 and concludes the grouping of all snippet
config classes.
